Add formatter module dir and move formatters to there
