+++ lib/ec2spec/formatter/slack_formatter.rb
@@ -1,0 +1,13 @@
+require 'json'
+
+module Ec2spec
+  module Formatter
+    module SlackFormatter
+      include Ec2spec::Formatter::PlainTextFormatter
+
+      def output(results, _hosts)
+        "```\n#{super}\n```"
+      end
+    end
+  end
+end
