--- lib/ec2spec/hash_formatter.rb
@@ -1,11 +1,0 @@
-require 'json'
-
-module Ec2spec
-  module HashFormatter
-    def output(results, _hosts)
-      results.each_with_object({}) do |result, hash|
-        hash[result.host] = result.to_hash
-      end
-    end
-  end
-end
