--- lib/ec2spec/slack_formatter.rb
@@ -1,11 +1,0 @@
-require 'json'
-
-module Ec2spec
-  module SlackFormatter
-    include Ec2spec::PlainTextFormatter
-
-    def output(results, _hosts)
-      "```\n#{super}\n```"
-    end
-  end
-end
